[REFACTOR] #47 - 온보딩 관련 뷰 패키지 이동
- ui/onboarding/view 로 패키지 이동
