+++ app/src/main/java/com/fork/spoonfeed/presentation/ui/onboarding/view/OnboardingActivity.kt
@@ -1,4 +1,4 @@
-package com.fork.spoonfeed.presentation.ui.onboarding
+package com.fork.spoonfeed.presentation.ui.onboarding.view
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,7 +6,7 @@ import com.fork.spoonfeed.R
 import com.fork.spoonfeed.databinding.ActivityOnboardingBinding
 import com.fork.spoonfeed.presentation.MainActivity
 import com.fork.spoonfeed.presentation.base.BaseViewUtil
-import com.fork.spoonfeed.presentation.ui.onboarding.signup.SignupActivity
+import com.fork.spoonfeed.presentation.ui.onboarding.view.signup.SignupActivity
 
 class OnboardingActivity :
     BaseViewUtil.BaseAppCompatActivity<ActivityOnboardingBinding>(R.layout.activity_onboarding) {
