+++ app/src/main/java/com/fork/spoonfeed/presentation/ui/onboarding/view/signup/SignupFinishFragment.kt
@@ -1,11 +1,8 @@
-package com.fork.spoonfeed.presentation.ui.onboarding.signup
+package com.fork.spoonfeed.presentation.ui.onboarding.view.signup
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import com.fork.spoonfeed.R
 import com.fork.spoonfeed.databinding.FragmentSignupFinishBinding
 import com.fork.spoonfeed.presentation.MainActivity
