+++ app/src/main/res/layout/fragment_signup_finish.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".presentation.ui.onboarding.signup.SignupFinishFragment">
+        tools:context=".presentation.ui.onboarding.view.signup.SignupFinishFragment">
 
         <ImageView
             android:id="@+id/iv_signup_finish"
